removed duplicated quicktreegrid
