--- src/components/QuickTreeGrid/QuickTreeGrid.scss
@@ -1,259 +1,0 @@
-@import '../../index.scss';
-
-.grid-component-container {
-   height: calc(100% - 40px);
-   width: calc(100% - 35px);
-   margin: 20px;
-
-   .group-drop-toolbar {
-       width: 100%;
-       box-sizing: border-box;
-       min-height: 42px;
-       padding: 2px;
-       border-bottom: 1px solid $secondary-color-20;
-       background-color: $disabled-field;
-
-        .group-drop-empty {
-            display: block;
-            width: 200px;
-            padding-left: 20px;
-            padding-top: 9px;
-            color: $secondary-color-45;
-        }
-
-        .group-column-boxes {
-            width: calc(100% - 65px);
-
-            .group-column-box {
-                width: 150px;
-                padding-left: 5px;
-                position: relative;
-                top: 3px;
-                display: inline-block;
-    
-                .group-text-boarder {
-                    box-sizing: border-box;
-                    border: 1px solid $secondary-color-20;
-                    border-radius: 8px;
-                    background-color: $white-color;
-                    cursor: pointer;
-                    height: 32px;
-                    line-height: 30px; // height minus the border
-                    position: relative;
-                    padding: 0 10px;
-                    display: block;
-    
-                    .header-text {
-                        span {
-                            overflow: hidden; 
-                            white-space: nowrap;
-                            text-overflow: ellipsis;
-                            width: 75%;
-                            display: inline-block;
-                            color: $primary-color;
-                            font-weight: 500;
-                        }
-                    }
-                }
-    
-                .icon-sort {
-                    top: 1px;
-                    right: 30px;
-                }
-    
-                .icon-remove-group {
-                    font-size: 12px;
-                    margin: 0;
-                    position: absolute;
-                    top: 0;
-                    right: 10px;
-                }
-            }
-        }
-
-        .expand-collapse-buttons {
-            float: right;
-            height: 100%;
-            position: absolute;
-            top: 12px;
-            right: 6px;
-
-            .icon {
-                cursor: pointer;
-                margin-right: 15px;
-            }
-        }
-    }
-
-    .icon-sort {
-        font-size: 8px;
-        margin: 0;
-        position: absolute;
-        top: 10px;
-        right: 5px;
-    }
-
-    .icon:hover {
-        color: $primary-color;
-    }
-
-    .grid-component-header {
-        overflow: hidden !important;
-    }
-
-    .grid-header-column {
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        border-bottom: 2px solid $secondary-color-20;
-        display: inline-block;
-        box-sizing: border-box;
-
-        .header-column-content {
-            display: inline-block;
-            width: calc(100% - 7px);
-
-            .header-text {
-                overflow: hidden;
-                text-overflow: ellipsis;
-                margin-right: 12px;
-                padding-top: 3px;
-
-                span {
-                    padding-left: 5px;
-                    font-weight: 500;
-                }
-            }
-
-            &.header-column-sortable {
-                cursor: pointer;
-            }
-        }
-
-        .grid-column-draggable {
-            width: 5px;
-            cursor: col-resize;
-            height: 100%;
-            float: right;
-            padding: 0 1px;
-            border-right: 1px solid $secondary-color-20;
-            box-sizing: border-box;
-        }
-    }
-
-    .grid-component {
-        .grid-component-cell {
-            user-select: none;
-
-            > div {
-                white-space: nowrap;
-                overflow: hidden;
-                text-overflow: ellipsis;
-                cursor: default;
-                padding-top: 5px;
-            }
-
-            &.is-hover {
-                background-color: $disabled-field;
-            }
-
-            &.is-selected {
-                background-color: $disabled-field;
-            }
-
-            .status-critical {
-                color: $error-status-server;
-            }
-            .status-warning {
-                color: $warning-status-server;
-            }
-            .status-ok {
-                color: $ok-status-server;
-            }
-
-            .dropdown {
-                padding-left: 5px;
-                
-                .dropdown-title {
-                    .icon {
-                        cursor: pointer;
-                    }
-                }
-            }
-
-            &.border-column-cell {
-                border-right: 1px solid $secondary-color-20;
-                box-sizing: border-box;
-            }
-        }
-
-        &.no-scrollbar {
-            overflow-x: hidden !important;
-        }
-
-        .grid-group-row {
-            font-weight: 500;
-
-            &.is-expanded {
-                border-bottom: 1px solid $secondary-color-20;
-            }
-
-            &.is-collapsed {
-                border-bottom: 0;
-            }
-
-            .grid-group-row-inner {
-                overflow: hidden;
-                text-overflow: ellipsis;
-                padding-top: 4px;
-
-                .icon {
-                    font-size: 10px;
-                    margin: 5px 0 0 0;
-                    position: absolute;
-                    cursor: pointer;
-                }
-
-                .group-row-text {
-                    padding-left: 7px;
-                    display: inline-block;
-                    cursor: pointer;
-
-                    &:hover {
-                        .icon {
-                            color: $primary-color;
-                        }
-                    }
-
-                    span {
-                        white-space: nowrap;
-                        overflow: hidden;
-                        text-overflow: ellipsis;
-                        color: $primary-color;
-                        padding-left: 15px;
-                    }
-                }
-            }
-        }
-    }
-
-    .grid-column-footer {
-        overflow-y: hidden !important;
-
-        .grid-footer-cell {
-            box-sizing: border-box;
-            border-top: 2px solid $secondary-color-20;
-            padding-top: 5px;
-            overflow: hidden;
-            white-space: nowrap;
-            text-overflow: ellipsis;
-            padding-left: 5px;
-            user-select: none;
-            cursor: default;
-        }
-    }
-}
-
-.expand-colapse-action-icon {
-    cursor: pointer;
-}
